Organises uploader code much better
