+++ app/config.py
@@ -1,0 +1,4 @@
+from transactions.db import MockTransactionDB
+
+db=MockTransactionDB()
+
